requirements.txt
aioice==0.9.0
aiortc==1.6.0
anyio==3.7.1
asgiref==3.6.0
av==11.0.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
click==8.1.7
contourpy==1.0.7
cryptography==41.0.7
cycler==0.11.0
dlib==19.24.2
dnspython==2.4.2
email-validator==1.3.1
exceptiongroup==1.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.108.0
fonttools==4.38.0
google-crc32c==1.5.0
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.4
ifaddr==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.6.3
numpy==1.24.1
orjson==3.9.10
packaging==23.0
pika==1.3.2
Pillow==9.4.0
pycparser==2.21
pydantic==1.10.13
pyee==11.1.0
pylibsrtp==0.9.1
pyOpenSSL==23.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
PyYAML==6.0.1
requests==2.28.2
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.32.0.post1
threadpoolctl==3.2.0
typing_extensions==4.9.0
ujson==5.9.0
urllib3==1.26.14
uvicorn==0.17.6
uvloop==0.19.0
watchgod==0.8.2
websockets==12.0
